clean:formatが異なるだけのls_shortとls_longのクラスを統合
