+++ 07.ls_object/ls_command.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
+require 'io/console'
 require_relative 'file_entries_list'
+require_relative 'dir_entry'
+
+SHORT_FORMAT_WHOLE_WIDTH = IO.console.winsize[1]
 
 class LsCommand
   def initialize(paths, options)
@@ -10,10 +14,26 @@ class LsCommand
   end
 
   def run
-    raise NotImplementedError, 'This method should be overridden by subclasses'
+    if @options[:long_format]
+      run_ls_long
+    else
+      run_ls_short
+    end
   end
 
   private
+
+  def run_ls_short
+    render_files_short unless @file_entries.empty?
+    puts if @file_entries.size.positive? && @dir_entries.size.positive?
+    render_directories_short unless @dir_entries.empty?
+  end
+
+  def run_ls_long
+    render_files_long unless @file_entries.empty?
+    puts if @file_entries.size.positive? && @dir_entries.size.positive?
+    render_directories_long unless @dir_entries.empty?
+  end
 
   def split_files_and_dirs(paths)
     files = []
@@ -29,15 +49,29 @@ class LsCommand
     [files, sorted_dirs]
   end
 
-  def render_files
-    raise NotImplementedError, 'This method should be overridden by subclasses'
+  def render_files_short
+    file_list = FileEntriesList.new(@file_entries, @options)
+    puts file_list.run_ls_short(SHORT_FORMAT_WHOLE_WIDTH)
   end
 
-  def render_directories
-    raise NotImplementedError, 'This method should be overridden by subclasses'
+  def render_directories_short
+    @dir_entries.each do |dir_entry|
+      puts "#{dir_entry.path}:" if dir_entry.path.directory? && (@dir_entries.size >= 2 || !@file_entries.empty?)
+      puts dir_entry.run_ls_short(SHORT_FORMAT_WHOLE_WIDTH)
+      puts if @dir_entries.size > 1
+    end
   end
 
-  def render
-    raise NotImplementedError, 'This method should be overridden by subclasses'
+  def render_files_long
+    file_list = FileEntriesList.new(@file_entries, @options)
+    puts file_list.run_ls_long
+  end
+
+  def render_directories_long
+    @dir_entries.each do |dir_entry|
+      puts "#{dir_entry.path}:" if @dir_entries.size > 1 || @file_entries.size.positive?
+      puts dir_entry.run_ls_long
+      puts if @dir_entries.size > 1
+    end
   end
 end
